+++ beanie/odm/settings/collection.py
@@ -1,8 +1,11 @@
-from typing import List, Type
+from typing import List, Type, Optional
 
 from motor.motor_asyncio import AsyncIOMotorCollection, AsyncIOMotorDatabase
 from pydantic.main import BaseModel
 from pymongo import IndexModel
+
+from beanie.exceptions import MongoDBVersionError
+from beanie.odm.settings.timeseries import TimeSeriesConfig
 
 
 class IndexModelField(IndexModel):
@@ -21,6 +24,7 @@ class IndexModelField(IndexModel):
 class CollectionInputParameters(BaseModel):
     name: str = ""
     indexes: List[IndexModelField] = []
+    timeseries: Optional[TimeSeriesConfig]
 
     class Config:
         arbitrary_types_allowed = True
@@ -63,8 +67,30 @@ class CollectionSettings(BaseModel):
         if not collection_parameters.name:
             collection_parameters.name = document_model.__name__
 
+        # check mongodb version
+        build_info = await database.command({"buildInfo": 1})
+        mongo_version = build_info["version"]
+        major_version = int(mongo_version.split(".")[0])
+
+        if collection_parameters.timeseries is not None and major_version < 5:
+            raise MongoDBVersionError(
+                "Timeseries are supported by MongoDB version 5 and higher"
+            )
+
         # create motor collection
-        collection = database[collection_parameters.name]
+        if (
+            collection_parameters.timeseries is not None
+            and collection_parameters.name
+            not in await database.list_collection_names()
+        ):
+
+            collection = await database.create_collection(
+                **collection_parameters.timeseries.build_query(
+                    collection_parameters.name
+                )
+            )
+        else:
+            collection = database[collection_parameters.name]
 
         # indexes
         old_indexes = (await collection.index_information()).keys()