+++ tests/odm/models.py
@@ -21,6 +21,7 @@ from pymongo import IndexModel
 from beanie import Document, Indexed, Insert, Replace, ValidateOnSave
 from beanie.odm.actions import before_event, after_event
 from beanie.odm.fields import Link
+from beanie.odm.settings.timeseries import TimeSeriesConfig
 
 
 class Option2(BaseModel):
@@ -45,8 +46,8 @@ class GeoObject(BaseModel):
 
 class Sample(Document):
     timestamp: datetime.datetime
-    increment: int
-    integer: int
+    increment: Indexed(int)
+    integer: Indexed(int)
     float_num: float
     string: str
     nested: Nested
@@ -284,10 +285,17 @@ class House(Document):
     windows: List[Link[Window]]
     door: Link[Door]
     roof: Optional[Link[Roof]]
-    name: str = Field(hidden=True)
-    height: int = 2
+    name: Indexed(str) = Field(hidden=True)
+    height: Indexed(int) = 2
 
 
 class DocumentForEncodingTest(Document):
     bytes_field: Optional[bytes]
     datetime_field: Optional[datetime.datetime]
+
+
+class DocumentWithTimeseries(Document):
+    ts: datetime.datetime = Field(default_factory=datetime.datetime.now)
+
+    class Collection:
+        timeseries = TimeSeriesConfig(time_field="ts", expire_after_seconds=2)